# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for BAST Request System.",
        "title": "BAST Request API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/bast-requests": {
            "get": {
                "description": "Retrieve a list of BAST requests, optionally filtered by customer_id, project_id, and status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-requests"
                ],
                "summary": "Get all BAST requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID filter",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Project ID filter",
                        "name": "project_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bast-request_internal_models.BastRequest"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Submit a new BAST request and automatically generate running number or use PO number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-requests"
                ],
                "summary": "Create a new BAST request",
                "parameters": [
                    {
                        "description": "Request Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.BastRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.BastRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/bast-requests/{id}": {
            "get": {
                "description": "Retrieve a single BAST request based on its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-requests"
                ],
                "summary": "Get a BAST request by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.BastRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/bast-requests/{id}/audit": {
            "get": {
                "description": "Retrieve the history of changes for a specific request ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-requests"
                ],
                "summary": "Get audit log for a request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/bast-requests/{id}/status": {
            "patch": {
                "description": "Update the status of a specific BAST request (Active, Used, Void)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-requests"
                ],
                "summary": "Update request status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status Data",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.UpdateStatusInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bast-request_internal_models.BastFormat": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "formatID": {
                    "type": "string"
                },
                "formatName": {
                    "type": "string"
                },
                "formatPattern": {
                    "description": "e.g. BAST/{YYYY}/{MM}/{SEQ}",
                    "type": "string"
                },
                "formatType": {
                    "description": "PO, Internal",
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "bast-request_internal_models.BastRequest": {
            "type": "object",
            "properties": {
                "bastNumber": {
                    "type": "string"
                },
                "bastRequestID": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/bast-request_internal_models.Customer"
                },
                "customerID": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "format": {
                    "$ref": "#/definitions/bast-request_internal_models.BastFormat"
                },
                "formatID": {
                    "type": "string"
                },
                "perihal": {
                    "type": "string"
                },
                "poNumber": {
                    "type": "string"
                },
                "project": {
                    "$ref": "#/definitions/bast-request_internal_models.Project"
                },
                "projectID": {
                    "type": "string"
                },
                "requestedAt": {
                    "type": "string"
                },
                "requestedBy": {
                    "type": "string"
                },
                "status": {
                    "description": "Active, Used, Void",
                    "type": "string"
                },
                "tipeNomor": {
                    "description": "PO, Internal",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "bast-request_internal_models.Customer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "customerCode": {
                    "type": "string"
                },
                "customerID": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "status": {
                    "description": "active, inactive",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "bast-request_internal_models.Project": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/bast-request_internal_models.Customer"
                },
                "customerID": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "projectCode": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "projectName": {
                    "type": "string"
                },
                "status": {
                    "description": "active, inactive",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_handlers.UpdateStatusInput": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}